pyproject.toml
[project]
name = "final-assignment-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio[oauth]>=5.37.0",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.28",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pygraphviz>=1.14",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "smolagents[litellm]>=0.5.0",
    "tavily-python>=0.7.9",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
